This Android Activity manages the user interface and logic for a To-Do list, handling task display, addition, and deletion.
It utilizes an SQLite database (TaskDatabaseHelper) to persistently store tasks along with their scheduled dates.

The class features methods to load and distinguish between today's tasks and future tasks and includes a date picker for task scheduling.

It incorporates a background thread to continuously update a TextView with the current date and time in real-time